+++ teamscale-gradle-plugin/src/test/kotlin/com/teamscale/TeamscalePluginTest.kt
@@ -129,31 +129,6 @@ class TeamscalePluginTest {
             .hasSize(19)
     }
 
-    @Test
-    fun `report directory is cleaned`() {
-        val oldReportFile = File(
-            temporaryFolder.root,
-            "build/reports/testwise-coverage/unitTest/old-report-Unit-Tests.json"
-        )
-        oldReportFile.parentFile.mkdirs()
-        oldReportFile.writeText("Test")
-        assertThat(oldReportFile).exists()
-
-        val build = build(
-            true, false,
-            "unitTest",
-            "--impacted",
-            "--run-all-tests",
-            "-PexcludeFailingTests=true"
-        )
-        assertThat(build.output).contains("SUCCESS (18 tests, 12 successes, 0 failures, 6 skipped)")
-            .doesNotContain("you did not provide all relevant class files")
-        val testwiseCoverageReportFile =
-            File(temporaryFolder.root, "build/reports/testwise-coverage/unitTest/Unit-Tests.json")
-        assertThat(testwiseCoverageReportFile).exists()
-        assertThat(oldReportFile).doesNotExist()
-    }
-
     private fun build(executesTask: Boolean, expectFailure: Boolean, vararg arguments: String): BuildResult {
         val runnerArgs = arguments.toMutableList()
         val runner = GradleRunner.create()